+++ lib/nosedrum/message_cache.ex
@@ -8,16 +8,7 @@ defmodule Nosedrum.MessageCache do
   after comparison of edited messages.
   """
 
-  alias Nostrum.Struct.{Channel, Guild, Message, User}
-
-  @typedoc """
-  A slimmed down message with the most relevant data.
-
-  You can fetch information about the author and the channel using
-  Nostrum's built in cache, specificaly `Nostrum.Cache.ChannelCache.get/1`
-  and `Nostrum.Cache.UserCache.get/1`.
-  """
-  @type cache_message :: {Message.id(), Channel.id(), User.id(), Message.content()}
+  alias Nostrum.Struct.{Guild, Message}
 
   @doc """
   Retrieve a single message for the given `guild_id` with the specified `message_id`
@@ -26,7 +17,7 @@ defmodule Nosedrum.MessageCache do
   If no message with the given ID is cached, `nil` is returned.
   """
   @callback get(guild_id :: Guild.id(), message_id :: Message.id(), cache :: reference()) ::
-              cache_message | nil | {:error, String.t()}
+              Message.t() | nil | {:error, String.t()}
 
   @doc """
   Retrieve up to `limit` messages for the given `guild_id` from the cache
@@ -40,19 +31,19 @@ defmodule Nosedrum.MessageCache do
               limit :: pos_integer | nil,
               cache :: reference()
             ) ::
-              {:ok, [cache_message]} | {:error, String.t()}
+              {:ok, [Message.t()]} | {:error, String.t()}
 
   @doc """
   Consume the given `message` in the cache process identified by `cache`.
 
   Whether this is done asynchronously or synchronously depends on the implementation.
   """
-  @callback consume(message :: Message.t(), cache :: reference()) :: :ok
+  @callback consume(message :: Message.t(), cache :: reference()) :: any()
 
   @doc """
   Update the given `message` in the cache, if present.
 
   Whether this is done asynchronously or synchronously depends on the implementation.
   """
-  @callback update(message :: Message.t(), cache :: reference()) :: :ok
+  @callback update(message :: Message.t(), cache :: reference()) :: any()
 end